resm0 & resm0-var now call residue-by-methods
This is a bit inefficient, I suppose. Maxima has determined the pole order, but this code sends n/e to residue-by-methods. But replacing these functions is an effort to replace all of the old residue code to use the new. We get a much nicer value for integrate(1/(x^4+3 x^2 + 1)^2,x,0,inf)